refactor: change price from double to BigDecimal

+++ apps/inventory-api/src/main/java/com/melo/inventory/model/Product.java
@@ -4,6 +4,7 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import java.math.BigDecimal;
 
 @Entity
 public class Product {
@@ -11,15 +12,15 @@ public class Product {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String name;
-    private double price;
+    private BigDecimal price;
 
     public Product() {}
 
     public Long getId(){return id;}
     public String getName(){return name;}
-    public double getPrice(){return price;}
+    public BigDecimal getPrice(){return price;}
 
     public void setId(Long id){this.id = id;}
     public void setName(String name){this.name = name;}
-    public void setPrice(double price){this.price = price;}
+    public void setPrice(BigDecimal price){this.price = price;}
 }
