+++ apps/inventory-api/src/main/java/com/melo/inventory/model/ProductRequest.java
@@ -3,22 +3,24 @@ package com.melo.inventory.model;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 
 public class ProductRequest {
     @NotBlank
     private String name;
 
     @Positive
-    private double price;
+    private BigDecimal price;
 
-    public ProductRequest(String name, double price){
+    public ProductRequest(String name, BigDecimal price){
         this.name = name;
         this.price = price;
     }
 
     public String getName(){return name;}
-    public double getPrice(){return price;}
+    public BigDecimal getPrice(){return price;}
 
     public void setName(String name){this.name = name;}
-    public void setPrice(double price){this.price = price;}
+    public void setPrice(BigDecimal price){this.price = price;}
 }
